refactor(qwen35): split TP response protocol into tp_executor/responses.rs
Move TpWorkerReply/TpWorkerResponse + the response-collection/validation functions
(recv_runtime_responses/collect_runtime_responses/validate_*_responses/reply_name/
recv_runtime_response) out of tp_executor.rs into a flat tp_executor/responses.rs
submodule. Item/struct-field visibility pub(super); root re-exports via use
responses::*. cargo check + fmt + test --lib 95/95 + tp2 7/7 (weights + 2 GPU + NCCL)
green.
